Fixed warnings

(cherry picked from commit 5ba4a1922e38571ed6c42b5e6389351f5a62aa7d)

# Conflicts:
#	ACE/ace/Arg_Shifter.cpp
#	ACE/ace/Arg_Shifter.h
#	ACE/ace/Based_Pointer_T.inl
#	ACE/ace/Netlink_Addr.inl
#	ACE/ace/OS_NS_stdlib.cpp
#	ACE/ace/OS_NS_time.inl
#	ACE/ace/Signal.inl
#	ACE/ace/Timer_Queue_T.cpp
#	ACE/ace/XML_Utils/XMLSchema/id_map.hpp
#	TAO/TAO_IDL/driver/drv_preproc.cpp

+++ ACE/ace/Log_Msg.inl
@@ -19,7 +19,7 @@ ACE_Log_Msg::log_priority_enabled (ACE_Log_Priority log_priority)
 {
   return ACE_BIT_ENABLED (this->priority_mask_ |
                           ACE_Log_Msg::process_priority_mask_,
-                          log_priority);
+                          static_cast<u_long> (log_priority));
 }
 
 ACE_INLINE
